# src/main/java/com/experianhealth/ciam/scimapi/entity/Name.java
package com.experianhealth.ciam.scimapi.entity;

import com.fasterxml.jackson.annotation.JsonInclude;

@JsonInclude(JsonInclude.Include.NON_NULL)
public class Name {

    private String givenName;
    private String familyName;
    private String formatted;

    public Name() {

    }

    public Name(String givenName, String familyName, String formatted) {
        super();
        this.givenName = givenName;
        this.familyName = familyName;
        this.formatted = formatted;
    }

    public String getGivenName() {
        return givenName;
    }

    public void setGivenName(String givenName) {
        this.givenName = givenName;
    }

    public String getFamilyName() {
        return familyName;
    }

    public void setFamilyName(String familyName) {
        this.familyName = familyName;
    }


    public String getFormatted() {
        return formatted;
    }

    public void setFormatted(String formatted) {
        this.formatted = formatted;
    }


}